Target the homepage rather than user status page.
The user status page is flagged on the target server for too many GETs
so the home page is used instead.  But the entire homepage isn't needed
so the ajax call is aborted after the first 1500 bytes if the progress
event is fired.

A regexp is used to search the resulting text for the karma value.

HTTPS is used instead of http.  This avoids the imminent redirect and
also allows the karma value to show up in the home page data.  If the
user is not logged in then the button just shows text 'login'.